% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_adjust}
\alias{spatial_adjust}
\title{spatial_adjust}
\usage{
spatial_adjust(
  object,
  slice_id,
  proj_name = NULL,
  rotate_mirror = c(0, "N"),
  feature = NULL,
  operator = NULL,
  cols = NULL,
  plot = T,
  verbose = T,
  pt_cex = 0.5
)
}
\arguments{
\item{object}{S4, slice_id character, proj_name character, rotate_mirror vector, feature character, operator character, cols vector, plot boolean, verbose boolean, pt_cex numeric}
}
\value{
list
}
\description{
creating a spatial object with digital matrix and showing spatial map with specified feature
}
