% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_binstat}
\alias{spatial_binstat}
\title{spatial_binstat}
\usage{
spatial_binstat(
  spatial_adjust_obj,
  bins = c(8, 8),
  min.cutoff = 0,
  min.cnt = 10,
  pct.cutoff = 0.5,
  bin_border_col = "gray",
  bin_target_col = "red",
  bin_border_lwd = 0.5,
  bin_target_lwd = 1,
  bin_border_lty = 3,
  bin_target_lty = 3,
  do.stat = T,
  mtx_msk = NULL,
  plot = T,
  label = F,
  pt_cex = 0.5
)
}
\arguments{
\item{spatial_adjust_obj}{list, bins vector, min.cutoff numeric, min.cnt numeric, pct.cutoff numeric, bin_border_col character, bin_target_col character, bin_border_lwd numeric, bin_target_lwd numeric, bin_border_lty numeric, bin_target_lty numeric, do.stat boolean, mtx_msk matrix, plot boolean, label boolean, pt_cex numeric}
}
\value{
list
}
\description{
creating binary matrix with specified feature and calculating significance of correlation between targeted feature and specified feature
}
