% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_magPlot}
\alias{spatial_magPlot}
\title{spatial_magPlot}
\usage{
spatial_magPlot(
  spaVecObj,
  cols = c("green", "orange"),
  cex_point = 1,
  cex_text = 1,
  cex_line = 1,
  lty = 1,
  ...
)
}
\arguments{
\item{spaVecObj}{list, a spatial_vector or spatial_vectorX object.}

\item{cols}{vector, a sequence of colors to use. Default, c('green','orange').}

\item{cex_point}{numeric, size of point. Default, 1.}

\item{cex_text}{numeric, size of text. Default, 1.}

\item{cex_line}{numeric, size of line. Default, 1.}

\item{lty}{numeric, line type. Default, 1.}
}
\description{
visualizing magnitude & projected score for spatial relatioship speculator
}
