% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_vecMagnitude}
\alias{spatial_vecMagnitude}
\title{spatial_vecMagnitude}
\usage{
spatial_vecMagnitude(spaVecObj)
}
\arguments{
\item{spaVecObj}{list, a spatial_vector or spatial_vectorX object.}
}
\value{
vector, a vector containing magnitude values for specified features.
}
\description{
calculating magnitude
}
