% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_vecPlot}
\alias{spatial_vecPlot}
\title{spatial_vecPlot}
\usage{
spatial_vecPlot(
  spaVecObj,
  cex_point = 1,
  cex_text = 1,
  pct_ctf = NULL,
  cols = NULL,
  projected = T,
  ...
)
}
\arguments{
\item{spaVecObj}{list, cex_point numeric, cex_text numeric, pct_ctf numeric, cols vector, projected boolean}
}
\description{
visualizing vectors for spatial relatioship speculator
}
