% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_vecProj}
\alias{spatial_vecProj}
\title{spatial_vecProj}
\usage{
spatial_vecProj(spaVecObj)
}
\arguments{
\item{spaVecObj}{list, a spatial_vector or spatial_vectorX object.}
}
\value{
vector, a vector containing projected scores for all specified features.
}
\description{
calculating projected score
}
