% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_vector}
\alias{spatial_vector}
\title{spatial_vector}
\usage{
spatial_vector(
  object,
  target,
  features,
  objNM = "obj",
  steps = seq(2, 12, 2),
  min.cutoffs = NULL,
  threads = NULL,
  stat = TRUE,
  package.libpath = NULL
)
}
\arguments{
\item{object}{S4, target character, features vector, objNM character, steps vector, min.cutoffs numeric, threads numeric, stat boolean, package.libpath character}
}
\value{
list
}
\description{
generating vectors for spatial relatioship speculator in one sample
}
