% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_vectorX}
\alias{spatial_vectorX}
\title{spatial_vectorX}
\usage{
spatial_vectorX(
  objects,
  target,
  features,
  steps = seq(2, 12, 2),
  min.cutoffs = NULL,
  threads = NULL,
  package.libpath = NULL,
  output2RDS = NULL
)
}
\arguments{
\item{objects}{list, target character, features vector, steps vector, min.cutoffs numeric, threads numeric, package.libpath character, output2RDS character}
}
\value{
list
}
\description{
generating vectors for spatial relatioship speculator across various samples
}
