% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_binstat}
\alias{spatial_binstat}
\title{spatial_binstat}
\usage{
spatial_binstat(
  spatial_adjust_obj,
  bins = c(8, 8),
  min.cutoff = 0,
  min.cnt = 10,
  pct.cutoff = 0.5,
  bin_border_col = "gray",
  bin_target_col = "red",
  bin_border_lwd = 0.5,
  bin_target_lwd = 1,
  bin_border_lty = 3,
  bin_target_lty = 3,
  do.stat = T,
  mtx_msk = NULL,
  plot = T,
  label = F,
  pt_cex = 0.5
)
}
\arguments{
\item{spatial_adjust_obj}{list,}

\item{bins}{vector,}

\item{min.cutoff}{numeric,}

\item{min.cnt}{numeric,}

\item{pct.cutoff}{numeric,}

\item{bin_border_col}{character,}

\item{bin_target_col}{character,}

\item{bin_border_lwd}{numeric,}

\item{bin_target_lwd}{numeric,}

\item{bin_border_lty}{numeric,}

\item{bin_target_lty}{numeric,}

\item{do.stat}{boolean,}

\item{mtx_msk}{matrix,}

\item{plot}{boolean,}

\item{label}{boolean,}

\item{pt_cex}{numeric}
}
\value{
list
}
\description{
creating binary matrix with specified feature
}
