% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_cordstat}
\alias{spatial_cordstat}
\title{spatial_cordstat}
\usage{
spatial_cordstat(
  spatial_adjust_obj1,
  spatial_adjust_obj2,
  min.cutoffs = c(0, 0),
  min.pct_cutoffs = NULL,
  min.cnt = c(10, 10),
  pct.cutoff = c(0.5, 0.5),
  operator_steps = c(4, 4, 4, 4),
  do.stat = T,
  plot = T,
  plot_bin = F,
  plot_bin_col = c("orange", "blue"),
  cord_mov = "add_X_add_Y",
  pt_cex = 0.5
)
}
\arguments{
\item{spatial_adjust_obj1}{list, spatial_adjust_obj2 list, min.cutoffs vector, min.pct_cutoffs vector, min.cnt vector, pct.cutoff vector, operator_steps vector, do.stat boolean, plot boolean, plot_bin boolean, plot_bin_col vector, cord_mov character, pt_cex numeric}
}
\value{
list
}
\description{
analyzing Jaccard indexes between two spatial_adjust objects with specified features and visualizing spatial map
}
