% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_magPlot}
\alias{spatial_magPlot}
\title{spatial_magPlot}
\usage{
spatial_magPlot(
  spaVecObj,
  cols = c("green", "orange"),
  cex_point = 1,
  cex_text = 1,
  cex_line = 1,
  lty = 1,
  ...
)
}
\arguments{
\item{spaVecObj}{list, cols vector, cex_point numeric, cex_text numeric, cex_line numeric, lty numeric}
}
\description{
visualizing magnitude & projected score for spatial relatioship speculator
}
