% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spatial_vecPlot}
\alias{spatial_vecPlot}
\title{spatial_vecPlot}
\usage{
spatial_vecPlot(
  spaVecObj,
  cex_point = 1,
  cex_text = 1,
  pct_ctf = NULL,
  cols = NULL,
  projected = T,
  ...
)
}
\arguments{
\item{spaVecObj}{list, a spatial_vector or spatial_vectorX object.}

\item{cex_point}{numeric, size of point. Default, 1.}

\item{cex_text}{numeric, size of text. Default, 1.}

\item{pct_ctf}{numeric, percentile cutoff for identifying features with extreme vector values, above which the feature name will be shown. Default, NULL.}

\item{cols}{vector, a sequence of colors to use. If not specified (NULL), a default color gradient from 'steelblue' to 'red' will be generated automatically.}

\item{projected}{boolean, determines whether the projetced score should be visualized. Default, TRUE.}
}
\description{
visualizing vectors for spatial relatioship speculator
}
